Move Styler protocol into its own file
